apparently seeds just gotta seed so we out here
